+++ crates/node/tests/globals_test.rs
@@ -1,4 +1,0 @@
-use proto_pdk_test_utils::*;
-
-#[cfg(not(windows))]
-generate_globals_test!("node-test", "prettier");